chore: :card_file_box: database changes and dependencies updating

+++ transactions/prisma/migrations/20250731230004_initial_config_v1/migration.sql
@@ -6,7 +6,7 @@ CREATE TABLE "public"."transactions" (
     "id" TEXT NOT NULL,
     "senderId" TEXT NOT NULL,
     "receiverId" TEXT NOT NULL,
-    "amount" BIGINT NOT NULL,
+    "amount" INTEGER NOT NULL,
     "description" TEXT,
     "status" "public"."TransactionStatus" NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
